requirements.txt
opencv-python
pytesseract
pillow
numpy
pydantic
python-dateutil
langid
pycountry
